break up components, fix test
